+++ tests/integration/cases/call_scalar_args/Java_heterogeneous_strategy_record_call@jdk_16.java
@@ -1,0 +1,8 @@
+class Main {
+static Object process(Object... args) { return null; }
+    public static void main() {
+process("hello");
+process(42);
+process(true);
+    }
+}